% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binsamp.R
\name{binsamp}
\alias{binsamp}
\title{Sampling fucntion}
\usage{
binsamp(iter, n, p)
}
\arguments{
\item{iter}{Number}

\item{n}{Number}

\item{p}{Number}
}
\description{
takes number of times to run the sample, iter, and samples n times each at p probability of success.
}
\examples{
binsamp(1000, 10, 0.5)

}
