% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap confidence interval fucntion}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number}

\item{x}{Vector}

\item{fun}{Text}

\item{alpha}{Decimal}

\item{cs}{Decimal}
}
\description{
takes number of iterations, a sample, a function, and an alpha and applies that function to the sample iter number of times and returns a bootstrap confidence interval.
}
\examples{
myboot2(x=sample)

}
