% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2pval.R
\name{myboot2pval}
\alias{myboot2pval}
\title{Bootstrap interval for a two sample test}
\usage{
myboot2pval(x1, x2, conf.level = 0.95, iter = 3000, mudiff = 0, test = "two")
}
\arguments{
\item{x1}{Vector}

\item{x2}{Vector}

\item{conf.level}{Number}

\item{iter}{Number}

\item{mudiff}{Number}

\item{test}{String}
}
\description{
takes two samples and returns the p-value and a bootstrap plot.
}
\examples{
set.seed(30);x=rnorm(15,mean=10,sd=4)
set.seed(40);y=rnorm(20,mean=12,sd=4)
myboot2pval(x,y)

}
