% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Central Limit Theorem for poisson distribution}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{Number}

\item{iter}{Number}

\item{lambda}{Number}
}
\description{
takes number of samples, n, and how many iterations, iter, for a Poisson Distribution with lambda.
}
\examples{
mycltp(n = 10, iter = 10000, lambda = 5)

}
