% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Plot Normal Curve and Probability}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Number}

\item{sigma}{Number}

\item{a}{Number}
}
\description{
takes mu and sigma and plots the Normal curve with a shaded area between the curve and x axis from
negative infinity to x=a. It also calculates the probability, P(X<=a)
}
\examples{
myncurve(3, 5, 4)

}
