% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise.R
\name{piecewise}
\alias{piecewise}
\title{Piecewise regression}
\usage{
piecewise(x, xk, coef)
}
\arguments{
\item{x}{Number}

\item{xk}{Number}

\item{coef}{Numeric Vector}
}
\value{
Numeric
}
\description{
takes the x value and plots a line until it hits the change point defined by xk,
then plots a different line to make a piecewise regression graph.
}
\examples{
curve(myf(x, xk, coef), add = TRUE, lwd = 2, col = "Blue")

}
